+++ packages/thumbprint-react/components/UIAction/themed.jsx
@@ -92,6 +92,10 @@ const Themed = React.forwardRef((props, ref) => {
                                 [styles.themedButtonThemeSecondary]: props.theme === 'secondary',
                                 [styles.themedButtonThemeCaution]: props.theme === 'caution',
                                 [styles.themedButtonThemeSolid]: props.theme === 'solid',
+                                [styles.themedButtonThemePopoverPrimary]:
+                                    props.theme === 'popover-primary',
+                                [styles.themedButtonThemePopoverSecondary]:
+                                    props.theme === 'popover-secondary',
                                 [styles.themedButtonWidthAuto]:
                                     props.width === 'auto' && !isWithinInputRow,
                                 [styles.themedButtonWidthFull]:
@@ -182,7 +186,15 @@ Themed.propTypes = {
     /**
      * Controls the button's background, text, and border color.
      */
-    theme: PropTypes.oneOf(['primary', 'secondary', 'tertiary', 'caution', 'solid']),
+    theme: PropTypes.oneOf([
+        'primary',
+        'secondary',
+        'tertiary',
+        'caution',
+        'solid',
+        'popover-primary',
+        'popover-secondary',
+    ]),
     /**
      * Changes the button's `line-height`, `padding`, `border-radius`, and `font-size`.
      */